Pull up following revision(s) (requested by rmind in ticket #829): 	usr.sbin/npf/npfctl/npfctl.8: revision 1.13 	usr.sbin/npf/npfctl/npf_build.c: revision 1.21 	lib/libnpf/npf.c: revision 1.18 	sys/net/npf/npf_ctl.c: revision 1.23 	usr.sbin/npf/npfctl/npfctl.h: revision 1.27 	lib/libnpf/npf.h: revision 1.15 	sys/net/npf/npf_ruleset.c: revision 1.19 	sys/net/npf/npf_impl.h: revision 1.28 	usr.sbin/npf/npfctl/npf_disassemble.c: revision 1.17 	usr.sbin/npf/npfctl/npfctl.c: revision 1.31 	usr.sbin/npf/npftest/libnpftest/npf_rule_test.c: revision 1.6 - Convert NPF dynamic rule ID to just incremented 64-bit counter. - Fix multiple bugs.  Also, update the man page.

+++ lib/libnpf/npf.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: npf.h,v 1.6.2.7 2013/02/11 21:49:48 riz Exp $	*/
+/*	$NetBSD: npf.h,v 1.6.2.8 2013/02/18 18:26:14 riz Exp $	*/
 
 /*-
  * Copyright (c) 2011-2013 The NetBSD Foundation, Inc.
@@ -79,8 +79,8 @@ void		npf_config_destroy(nl_config_t *);
 nl_config_t *	npf_config_retrieve(int, bool *, bool *);
 int		npf_config_flush(int);
 
-int		npf_ruleset_add(int, const char *, nl_rule_t *, uintptr_t *);
-int		npf_ruleset_remove(int, const char *, uintptr_t);
+int		npf_ruleset_add(int, const char *, nl_rule_t *, uint64_t *);
+int		npf_ruleset_remove(int, const char *, uint64_t);
 int		npf_ruleset_remkey(int, const char *, const void *, size_t);
 int		npf_ruleset_flush(int, const char *);
